pipelines/requirements.txt
kfp==1.7.0
google.cloud.aiplatform==1.8.0
